requirements.txt
# requirements.txt
tensorflow-macos
tensorflow-metal
keras
tensorboard
tensorflow-estimator

ml-dtypes==0.2.0
numpy>=1.22.0
opencv-python==4.7.0.72

scikit-learn>=1.3.0
scikit-image>=0.21.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
Pillow>=10.0.0
scipy>=1.11.0